change (int) to (float) for more exact calculations

+++ src/base/Provider.php
@@ -355,9 +355,9 @@ abstract class Provider extends SavableComponent implements ProviderInterface
         }
 
         return [
-            'length' => (int)$maxWidth,
-            'width'  => (int)$maxLength,
-            'height' => (int)$totalHeight,
+            'length' => (float)$maxWidth,
+            'width'  => (float)$maxLength,
+            'height' => (float)$totalHeight,
         ];
     }
 
@@ -383,9 +383,9 @@ abstract class Provider extends SavableComponent implements ProviderInterface
         $height = $orderHeight->toUnit($dimensionUnit);
 
         return [
-            'length' => (int)$length,
-            'width'  => (int)$width,
-            'height' => (int)$height,
+            'length' => (float)$length,
+            'width'  => (float)$width,
+            'height' => (float)$height,
             'weight' => (float)$weight,
         ];
     }
